# pkg/apigen/openapi/api.gen.go
// Package apigen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package apigen

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// CommonResponse Common response wrapper
type CommonResponse struct {
	Items []ResponseItem `json:"items"`
}

// ErrorResponse Error wrapper
type ErrorResponse struct {
	// Details Optional error details
	Details *map[string]interface{} `json:"details,omitempty"`

	// Error Error message
	Error string `json:"error"`
}

// ResponseItem defines model for ResponseItem.
type ResponseItem struct {
	CarouselImageUrls []string `json:"carouselImageUrls"`
	Description       string   `json:"description"`
	MainImageUrl      string   `json:"mainImageUrl"`
	Name              string   `json:"name"`
}

// TextImageRequest defines model for TextImageRequest.
type TextImageRequest struct {
	// Image URL to image
	Image string `json:"image"`

	// Text Input text
	Text string `json:"text"`
}

// TextRequest defines model for TextRequest.
type TextRequest struct {
	// Text Input text
	Text string `json:"text"`
}

// RespondTextImageJSONRequestBody defines body for RespondTextImage for application/json ContentType.
type RespondTextImageJSONRequestBody = TextImageRequest

// RespondTextJSONRequestBody defines body for RespondText for application/json ContentType.
type RespondTextJSONRequestBody = TextRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Respond to text and image input
	// (POST /api/v1/respond/image)
	RespondTextImage(ctx echo.Context) error
	// Respond to text input
	// (POST /api/v1/respond/text)
	RespondText(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// RespondTextImage converts echo context to params.
func (w *ServerInterfaceWrapper) RespondTextImage(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.RespondTextImage(ctx)
	return err
}

// RespondText converts echo context to params.
func (w *ServerInterfaceWrapper) RespondText(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.RespondText(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/api/v1/respond/image", wrapper.RespondTextImage)
	router.POST(baseURL+"/api/v1/respond/text", wrapper.RespondText)

}

type RespondTextImageRequestObject struct {
	Body *RespondTextImageJSONRequestBody
}

type RespondTextImageResponseObject interface {
	VisitRespondTextImageResponse(w http.ResponseWriter) error
}

type RespondTextImage200JSONResponse CommonResponse

func (response RespondTextImage200JSONResponse) VisitRespondTextImageResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type RespondTextImage400JSONResponse ErrorResponse

func (response RespondTextImage400JSONResponse) VisitRespondTextImageResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type RespondTextImage500JSONResponse ErrorResponse

func (response RespondTextImage500JSONResponse) VisitRespondTextImageResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type RespondTextRequestObject struct {
	Body *RespondTextJSONRequestBody
}

type RespondTextResponseObject interface {
	VisitRespondTextResponse(w http.ResponseWriter) error
}

type RespondText200JSONResponse CommonResponse

func (response RespondText200JSONResponse) VisitRespondTextResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type RespondText400JSONResponse ErrorResponse

func (response RespondText400JSONResponse) VisitRespondTextResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type RespondText500JSONResponse ErrorResponse

func (response RespondText500JSONResponse) VisitRespondTextResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Respond to text and image input
	// (POST /api/v1/respond/image)
	RespondTextImage(ctx context.Context, request RespondTextImageRequestObject) (RespondTextImageResponseObject, error)
	// Respond to text input
	// (POST /api/v1/respond/text)
	RespondText(ctx context.Context, request RespondTextRequestObject) (RespondTextResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// RespondTextImage operation middleware
func (sh *strictHandler) RespondTextImage(ctx echo.Context) error {
	var request RespondTextImageRequestObject

	var body RespondTextImageJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.RespondTextImage(ctx.Request().Context(), request.(RespondTextImageRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RespondTextImage")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(RespondTextImageResponseObject); ok {
		return validResponse.VisitRespondTextImageResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// RespondText operation middleware
func (sh *strictHandler) RespondText(ctx echo.Context) error {
	var request RespondTextRequestObject

	var body RespondTextJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.RespondText(ctx.Request().Context(), request.(RespondTextRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RespondText")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(RespondTextResponseObject); ok {
		return validResponse.VisitRespondTextResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}