# swagger/openapi.yml
openapi: 3.0.3
info:
  title: Pod API
  version: 1.0.0
  description: API for responding to text or text+image inputs
servers:
  - url: /
paths:
  /api/v1/respond/text:
    post:
      operationId: RespondText
      summary: Respond to text input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TextRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommonResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/respond/image:
    post:
      operationId: RespondTextImage
      summary: Respond to text and image input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TextImageRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommonResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ErrorResponse:
      type: object
      description: Error wrapper
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: object
          additionalProperties: true
          description: Optional error details
    TextRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Input text
    TextImageRequest:
      type: object
      required:
        - text
        - image
      properties:
        text:
          type: string
          description: Input text
        image:
          type: string
          format: url
          description: URL to image
    CommonResponse:
      type: object
      description: Common response wrapper
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ResponseItem"
    ResponseItem:
      type: object
      required:
        - name
        - description
        - mainImageUrl
        - carouselImageUrls
      properties:
        name:
          type: string
        description:
          type: string
        mainImageUrl:
          type: string
          format: uri
        carouselImageUrls:
          type: array
          items:
            type: string
            format: uri